BSP-287: Combined two separate common function classes, updated tests and relevant files
